reorg the gpu context and the unit tests
