+++ lmcache/v1/multiprocess/gpu_context.py
@@ -1,0 +1,191 @@
+# SPDX-License-Identifier: Apache-2.0
+"""
+GPU Cache Context management for LMCache multiprocessing.
+
+This module provides GPU-side KV cache management functionality, including:
+- GPUCacheContext: Manages shape and pointers to vLLM GPU KV cache tensors
+- Helper functions for tensor operations and key resolution
+"""
+
+# Standard
+import array
+
+# Third Party
+import cupy
+import torch
+
+# First Party
+from lmcache.logging import init_logger
+from lmcache.utils import _lmcache_nvtx_annotate
+from lmcache.v1.multiprocess.custom_types import (
+    KVCache,
+)
+
+logger = init_logger(__name__)
+
+
+def unwrap_kv_cache_tensors(kv_caches: KVCache) -> list[torch.Tensor]:
+    unwrapped_tensors = []
+    for ipc_wrapper in kv_caches:
+        tensor = ipc_wrapper.to_tensor()
+        unwrapped_tensors.append(tensor)
+    return unwrapped_tensors
+
+
+def list_to_gpu_tensor(lis: list[int], device: torch.device) -> torch.Tensor:
+    return torch.frombuffer(array.array("l", lis), dtype=torch.long).to(
+        device, non_blocking=True
+    )
+
+
+class GPUCacheContext:
+    """
+    Manages the shape and pointers to vLLM GPU KV cache tensors.
+    """
+
+    def __init__(self, kv_caches: KVCache, lmcache_chunk_size: int = 256):
+        self.kv_caches_ = unwrap_kv_cache_tensors(kv_caches)
+        self.device_ = self.kv_caches_[0].device
+
+        # Pointers
+        pointers_list = [t.data_ptr() for t in self.kv_caches_]
+        self.kv_cache_pointers_ = list_to_gpu_tensor(pointers_list, self.device_)
+
+        # MLA flag
+        # MLA shape: [num_blocks, block_size, hidden_dim]
+        # MHA shape: [2, num_blocks, block_size, num_heads, head_size]
+        self.is_mla_ = self.kv_caches_[0].ndim == 3
+
+        # Shape related
+        self.num_layers_ = len(self.kv_caches_)
+        if self.is_mla_:
+            self.num_blocks_ = self.kv_caches_[0].shape[0]
+            self.block_size_ = self.kv_caches_[0].shape[1]
+            self.hidden_dim_size_ = self.kv_caches_[0].shape[2]
+        else:
+            self.num_blocks_ = self.kv_caches_[0].shape[1]
+            self.block_size_ = self.kv_caches_[0].shape[2]
+            # hidden_dim = num_heads * head_size
+            num_heads = self.kv_caches_[0].shape[3]
+            head_size = self.kv_caches_[0].shape[4]
+            self.hidden_dim_size_ = num_heads * head_size
+
+        # Pre-computed slot mapping
+        # shape: [num_blocks, block_size]
+        block_ids = torch.arange(
+            0, self.num_blocks_, dtype=torch.long, device=self.device_
+        ).unsqueeze(1)
+        offsets = torch.arange(
+            0, self.block_size_, dtype=torch.long, device=self.device_
+        ).unsqueeze(0)
+        self.slot_mapping_tensor_ = (offsets + block_ids * self.block_size_).reshape(
+            (self.num_blocks, self.block_size_)
+        )
+
+        # Temporary GPU buffer for transfers
+        tmp_buffer_shape = self.get_kv_buffer_shape(lmcache_chunk_size)
+        self.tmp_gpu_buffer_ = torch.empty(
+            tmp_buffer_shape, dtype=self.dtype, device=self.device_
+        )
+
+        # Cuda streams
+        self.cuda_stream_ = torch.cuda.Stream(device=self.device_)
+        self.cupy_stream_ = cupy.cuda.ExternalStream(
+            self.cuda_stream_.cuda_stream, self.device_.index
+        )
+
+        # Extra initialization
+        self.cupy_stream_.launch_host_func(
+            lambda logger: logger.info(
+                "Initialized cuda stream on device %s", str(self.device_)
+            ),
+            logger,
+        )
+
+    @property
+    def dtype(self) -> torch.dtype:
+        return self.kv_caches_[0].dtype
+
+    @property
+    def device(self) -> torch.device:
+        return self.device_
+
+    @property
+    def kv_tensors(self) -> list[torch.Tensor]:
+        return self.kv_caches_
+
+    @property
+    def kv_pointers(self) -> torch.Tensor:
+        """
+        Returns a GPU tensor of the KV cache pointers
+        """
+        return self.kv_cache_pointers_
+
+    @property
+    def stream(self) -> torch.cuda.Stream:
+        """
+        Returns the CUDA stream for KV cache operations
+        """
+        return self.cuda_stream_
+
+    @property
+    def cupy_stream(self) -> cupy.cuda.Stream:
+        return self.cupy_stream_
+
+    @property
+    def block_size(self) -> int:
+        """
+        Returns the block size (number of tokens per block)
+        """
+        return self.block_size_
+
+    @property
+    def num_layers(self) -> int:
+        """
+        Returns the number of layers in the model
+        """
+        return self.num_layers_
+
+    @property
+    def num_blocks(self) -> int:
+        """
+        Returns the number of blocks in the KV cache
+        """
+        return self.num_blocks_
+
+    @property
+    def hidden_dim_size(self) -> int:
+        """
+        Returns the hidden dimension size of the model
+        """
+        return self.hidden_dim_size_
+
+    @property
+    def is_mla(self) -> bool:
+        """
+        Returns whether the model uses MLA
+        """
+        return self.is_mla_
+
+    def get_tmp_gpu_buffer(self, num_tokens: int) -> torch.Tensor:
+        """
+        Returns the temporary GPU buffer for transfers
+        """
+        return self.tmp_gpu_buffer_[:, :, :num_tokens, :]
+
+    @_lmcache_nvtx_annotate
+    def get_slot_mapping_tensor(self, gpu_block_ids: list[int]) -> torch.Tensor:
+        """
+        Returns the slot mapping tensor for the KV cache on GPU
+        """
+        gpu_block_ids_tensor = list_to_gpu_tensor(gpu_block_ids, self.device_)
+        return self.slot_mapping_tensor_[gpu_block_ids_tensor].flatten().contiguous()
+
+    def get_kv_buffer_shape(self, num_tokens: int) -> torch.Size:
+        """
+        Returns the shape of the KV buffer for the given number of tokens
+        """
+        if self.is_mla_:
+            return torch.Size((1, self.num_layers_, num_tokens, self.hidden_dim_size_))
+        else:
+            return torch.Size((2, self.num_layers_, num_tokens, self.hidden_dim_size_))
